% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qM.sM.R
\name{qM.sM}
\alias{qM.sM}
\title{Matrix Quantiles}
\usage{
qM.sM(sM, quantiles = c(0.025, 0.5, 0.975))
}
\arguments{
\item{sM}{an m x n x s array}

\item{quantiles}{quantiles to be computed}
}
\value{
an array of dimension m x n x l, where l is the length of
\code{quantiles}
}
\description{
Computes quantiles along the third dimension of a 3-d array.
}
\author{
Peter Hoff
}
\keyword{array}
\keyword{fit<-sbgcop.mcmc(swiss)}
\keyword{multivariate}
\keyword{qC<-qM.sM(fit$C.psamp)}
\keyword{qR<-qM.sM(sR.sC(fit$C.psamp))}
