% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwish.R
\name{rwish}
\alias{rwish}
\title{Sample from the Wishart Distribution}
\usage{
rwish(S0, nu)
}
\arguments{
\item{S0}{a positive definite matrix}

\item{nu}{a positive integer}
}
\value{
a positive definite matrix.
}
\description{
Generate a random sample from the Wishart distribution.
}
\details{
Return the sum of nu i.i.d.  rank-one matrices generated as \code{z\%*\%t(z)},
where \code{z} is a sample from a multivariate normal distribution with
covariance \code{S0}. The resulting random variable has mean \code{nu*S0}.
}
\author{
Peter Hoff
}
\keyword{datagen}
\keyword{distribution}
\keyword{multivariate}
