% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgcop-package.R
\docType{package}
\name{sbgcop-package}
\alias{sbgcop-package}
\alias{sbgcop}
\title{Semiparametric Bayesian Gaussian Copula Estimation and Imputation}
\description{
Estimation and inference for parameters in a Gaussian copula model, treating
univariate marginal distributions as nuisance parameters as described in
Hoff (2007) <doi:10.1214/07-AOAS107>. This pacakge also provides a 
semiparametric imputation procedure for
missing multivariate data.
}
\details{
\tabular{ll}{ Package: \tab sbgcop\cr Type: \tab Package\cr Version: \tab
0.980 \cr Date: \tab 2018-05-25\cr License: \tab GPL Version 2 or later \cr
} This function produces MCMC samples from the posterior distribution of a
correlation matrix, using a scaled inverse-Wishart prior distribution and an
extended rank likelihood. It also provides imputation for missing values in
a multivariate dataset.
}
\examples{


fit<-sbgcop.mcmc(swiss)
summary(fit)
plot(fit)


}
\references{
Hoff (2007) ``Extending the rank likelihood for semiparametric
copula estimation''
}
\author{
Peter Hoff <peter.hofff@duke.edu>
}
\keyword{multivariate}
