% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotci.sA.R
\name{plotci.sA}
\alias{plotci.sA}
\title{Plot Confidence Bands for Association Parameters}
\usage{
plotci.sA(sA, ylabs = colnames(sA[, , 1]), mgp = c(1.75, 0.75, 0))
}
\arguments{
\item{sA}{a p x p x nsamp array}

\item{ylabs}{a p x 1 vector of names for plotting labels}

\item{mgp}{margin parameters}
}
\description{
Plots 95% confidence bands based on an array of association parameters
}
\examples{

fit<-sbgcop.mcmc(swiss)

plotci.sA(fit$C.psamp) 

}
\author{
Peter Hoff
}
\keyword{array}
